Moving class ComplexSparseMatrix into complex_operator.?pp.
