+++ linalg/complex_operator.hpp
@@ -13,6 +13,7 @@
 #define MFEM_COMPLEX_OPERATOR
 
 #include "operator.hpp"
+#include "sparsemat.hpp"
 
 namespace mfem
 {
@@ -103,6 +104,30 @@ protected:
 
    mutable Vector x_r_, x_i_, y_r_, y_i_;
    mutable Vector *u_, *v_;
+};
+
+
+/** @brief Specialization of the ComplexOperator built from a pair of Sparse
+    Matrices.
+
+    The purpose of this specialization is to construct a single SparseMatrix
+    object which is equivalent to the 2x2 block system that the ComplexOperator
+    mimics. The resulting SparseMatrix can then be passed along to solvers which
+    require access to the CSR matrix data such as SuperLU, STRUMPACK, or similar
+    sparse linear solvers.
+
+    See ComplexOperator documentation in operator.hpp for more information.
+ */
+class ComplexSparseMatrix : public ComplexOperator
+{
+public:
+   ComplexSparseMatrix(SparseMatrix * A_Real, SparseMatrix * A_Imag,
+                       bool ownReal, bool ownImag,
+                       Convention convention = HERMITIAN)
+      : ComplexOperator(A_Real, A_Imag, ownReal, ownImag, convention)
+   {}
+
+   SparseMatrix * GetSystemMatrix() const;
 };
 
 }
