+++ linalg/sparsemat.hpp
@@ -18,7 +18,6 @@
 #include "../general/table.hpp"
 #include "../general/globals.hpp"
 #include "densemat.hpp"
-#include "complex_operator.hpp"
 
 namespace mfem
 {
@@ -628,29 +627,6 @@ template<> inline void Swap<SparseMatrix>(SparseMatrix &a, SparseMatrix &b)
 {
    a.Swap(b);
 }
-
-/** @brief Specialization of the ComplexOperator built from a pair of Sparse
-    Matrices.
-
-    The purpose of this specialization is to construct a single SparseMatrix
-    object which is equivalent to the 2x2 block system that the ComplexOperator
-    mimics. The resulting SparseMatrix can then be passed along to solvers which
-    require access to the CSR matrix data such as SuperLU, STRUMPACK, or similar
-    sparse linear solvers.
-
-    See ComplexOperator documentation in operator.hpp for more information.
- */
-class ComplexSparseMatrix : public ComplexOperator
-{
-public:
-   ComplexSparseMatrix(SparseMatrix * A_Real, SparseMatrix * A_Imag,
-                       bool ownReal, bool ownImag,
-                       Convention convention = HERMITIAN)
-      : ComplexOperator(A_Real, A_Imag, ownReal, ownImag, convention)
-   {}
-
-   SparseMatrix * GetSystemMatrix() const;
-};
 
 }
